requirements.txt
black==25.1.0
click==8.1.8
fastapi==0.128.0
ruff==0.15.0
freezegun==1.5.2
mutmut==3.3.0
pre-commit==4.5.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pytest==8.4.2
python-dotenv==1.1.0
requests==2.32.5
uvicorn>=0.30.0
google-generativeai>=0.8.0
rich>=13.0.0
openai>=1.0.0
reportlab>=4.0.0
playwright>=1.50.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
email-validator>=2.0.0
